refactor(pipeline): make testing/index.ts an exports-only barrel
Move captureSink/CaptureSink to capture.ts and scriptedSource to
scripted-source.ts; index.ts now only re-exports.
